mainpage-calculator 유효성 검사 로직을 별도의 함수로 분리
