+++ src/components/main/Calculator/Calculator.jsx
@@ -6,36 +6,8 @@ import label from 'assets/styles/common/label.module.css';
 import DoughnutChart from 'components/common/DoughnutChart/DoughnutChart';
 import { FcCalculator } from 'react-icons/fc';
 import { useState } from 'react';
+import { checkForm } from 'utils/helpers/calculator.js';
 
-// 유효성 검사
-const checkForm = (FormData) => {
-    const errors = {};
-
-    const { age, gender, height, weight } = FormData;
-
-    if (!age) {
-        errors.age = '나이를 입력해주세요';
-    }
-    if (!gender) {
-        errors.gender = '성별을 입력해주세요';
-    }
-    if (!height) {
-        errors.height = '신장을 입력해주세요';
-    } else {
-        if (height < 100 || height > 251) {
-            errors.height = '올바르지 않은 값입니다';
-        }
-    }
-    if (!weight) {
-        errors.weight = '체중을 입력해주세요';
-    } else {
-        if (weight < 20 || weight > 635) {
-            errors.weight = '올바르지 않은 값입니다';
-        }
-    }
-
-    return errors;
-};
 
 function Calculator() {
     // 입력한 사용자의 정보 (나이, 성별, 신장, 체중)
