+++ src/utils/helpers/calculator.js
@@ -25,8 +25,39 @@ const calcDay = (date) => {
     return Math.floor(day / (1000 * 60 * 60 * 24));
 }
 
+// mainpage-calculator 유효성 검사 함수
+const checkForm = (formData) => {
+    const errors = {};
+
+    const { age, gender, height, weight } = formData;
+
+    if (!age) {
+        errors.age = '나이를 입력해주세요';
+    }
+    if (!gender) {
+        errors.gender = '성별을 입력해주세요';
+    }
+    if (!height) {
+        errors.height = '신장을 입력해주세요';
+    } else {
+        if (height < 100 || height > 251) {
+            errors.height = '올바르지 않은 값입니다';
+        }
+    }
+    if (!weight) {
+        errors.weight = '체중을 입력해주세요';
+    } else {
+        if (weight < 20 || weight > 635) {
+            errors.weight = '올바르지 않은 값입니다';
+        }
+    }
+
+    return errors;
+}
+
 export {
     calcTotalCalorie,
     calcTotalPeriod,
     calcDay,
+    checkForm, 
 };
